Change the build command to not rely on environment variables in org.eclipse.cdt.core.prefs
Add parameters to the call to ```make``` which contain the CONFIG and TARGET definitions for each build.
These were part of the environment in which ```make``` was run by Eclipse but this needed the ```org.eclipse.cdt.core.prefs``` file in the ```.settings``` directory.
With this change an Eclipse build of the FT90x_Debug configuration will use the command line:
```
make CONFIG=Debug TARGET=ft900 all
```