Increase TDRST on downward facing hub port #30
This increase adds 10 ms to the time that enumeration waits for a reset driven on a downstream hub port. Previously the delay was for 10 ms, however Table 7-13 has 10 ms as the minimum time and 20 ms as the maximum time. Since this is controlled by the hub the hub may drive it for 20 ms. See "11.5.1.5 Resetting".