Test code for @NessDan. Vary S-Mask and C-Mask.
In one of the reported configurations in #30 there's a good chance that multiple full-speed (or low-speed) periodic schedule transactions are linked in the same branch in the periodic frame list. This will offset the start microframe for the transactions to each device. This allows the host a complete microframe to send the request to the each device rather than bunch multiple in the same microframe. They will still overlap although the complete frame mask span has been reduced from the original.
The host could place each device on it's own frame index entry since (in this configuration) there are 16 entries used that will produce the same result.